requirements.txt
streamlit
opencv-python
numpy
scikit-learn
Pillow
matplotlib
pandas
tensorflow